Reworked mapbrowser* things
